migration: discover the genesis block via super-oc tracking (#585)
Split from #502 

#### Problem
We need to actually discover the genesis block once the migration has
started in order to cast our BLS vote.
From the paper:
```
Definition 1 (strong Optimistic Confirmation). OCp(b) is the condition that
the child of b is in slot slot(b) + 1, and it includes the votes of at least p of stake
on b.
```
English: Block `b` whose child in the next slot contains votes for `b`
from 82% of stake is super OC

And to figure out the genesis block:
```
Definition 2 (look for genesis). If there exists an ancestor bOC of b such that
slot(bOC) ≥ s −1 and OCp(bOC) = true, then the function try(b) returns the
lowest ancestor b′of b such that slot(b′) < s.
```
English: From the super OC block, genesis is the greatest by slot #
ancestor from before the migration slot

#### Summary of Changes
Add tracking in collect vote lockouts and set the genesis block once
it's found.

Originally was going to do this in Cluster Info Vote Listener (CIVL)
because there was some confusion about including gossip votes or votes
which landed in children *not* in the next slot, but think it's clearer
and easier to debug if we only include replay votes as the paper
intended.