migration: Send the GenesisVote when we discover the genesis block (#589)
Split from #502 

#### Problem
Once compute bank stats discovers the genesis block we need to send out
the GenesisVote for it.

#### Summary of Changes
In replay if alpenglow is yet to be enabled and we have discovered a
genesis block, send the vote.
Refactored voting_utils to be accessible from replay by passing in
arguments rather than the VotingContext