fast leader handover: transmit optimistic parent to block creation loop (#581)
#### Problem and Summary of Changes
For fast leader handover, in addition to `votor` sending over a
confirmed parent thanks to a parent ready update, we need replay stage
to optimistically send block creation loop a parent to build on. This PR
gets replay stage to send the optimistic parent when the next slot is
the beginning of the validator's leader window.