block_component: Add GenesisCertificate block marker (#599)
#### Problem
We need some method to indicate the first Alpenglow block post migration

#### Summary of Changes
Add a block marker that will only be disseminated on the first alpenglow
block (direct child of genesis).
This block marker holds the `GenesisCertificate`.

Future PR will populate this in BCL and process it in replay to update
the genesis account.

Full details of migration in SIMD
https://github.com/solana-foundation/solana-improvement-documents/pull/384/files#diff-9a9498ea503301d23f7af15f620fcb53b060f666419c9214a3fe72ff465affd9R118