blockstore_processor: advance migration during startup (#608)
#### Problem
We currently do not advance `MigrationStatus` during startup at all,
it's just set at the bank forks root.
This causes problems if we restart after the migration has succeeded or
during the migrationary period before we've advanced the root.
The old root could indicate that we are still in the migration while
blockstore has already been cleaned up and we have Alpenglow blocks
present.

Additionally we do not process the `GenesisCertificate` marker in the
first Alpenglow block. This is key to inform startup and sometimes
lagging steady state that the migration was successful. There's no need
to try to do genesis discovery if this marker is processed, we can just
go straight into Alpenglow.

#### Summary of Changes
Allow processing of the `GenesisCertificate` marker:
- Immediately move us to `ReadyToEnable` as the processing of the
`GenesisCertificate` via block means we have the genesis block frozen
- Allow header and genesis certificate markers to be processed when
we're in the migrationary period, for such cases where we missed the
migration's completion and are receiving the first alpenglow block.

Advance `MigrationStatus` during startup in `load_frozen_forks`:
- If we root the feature flag activation, transition from
`PreFeatureActivation` to `Migration`
- Don't root startup blocks that are in the `Migration`
- If we process a `GenesisCertificate` marker go to `ReadyToEnable`
- If we are `ReadyToEnable`, enable alpenglow immediately and retry
processing the current block onwards as Alpenglow blocks instead
- Do not do genesis discovery - if we don't process a
`GenesisCertificate` then we rely on the standard `compute_bank_stats`
pathway in replay to find the genesis block post startup

Add a test where a node restarts post migration from a root before the
migration and ensure it can catch up.
Add a test where a node misses the migration entirely and ensure that it
can catch up.

Note: The certificate in the `GenesisCertificate` marker is not
validated yet. Future PR will plug in BLS verification.