feat: Introduces merkle tree struct (#640)
#### Problem

This is roughly a backport of
https://github.com/anza-xyz/agave/pull/9180.

Currently a merkle tree is represented as a simple list of nodes. We are
in the process of introducing double merkle for alpenglow and this
creates some problems for that.

In particular, we are going to have different types of merkle trees:
chain and double. It will be good to represent the different types of
trees using different types to improve readability of the code and to
enforce some checks. See:
https://github.com/anza-xyz/alpenglow/pull/518/files#diff-3c963fba687007333971f9a8049f494d9241e1b23693ee81da9c8f569aa1c376R29
for what we are envisioning.

#### Summary of Changes

This PR introduces a small initial change to represent a merkle tree
using a struct instead of a list of nodes. This has additional benefits
of removing some panics when we look up the root of the tree and that
the callers do not need to know precisely where in the list the root
lives.