refactor: reduces code duplication in bls_sigverifier (#632)
#### Problem

There is a lot of code duplication in the various tests of
bls_sigverifier as they all have to create channels to pass to the setup
code.


#### Summary of Changes

Reduces the code duplication by moving the channel creation into the
setup code.

This also makes it easier to add new channels to the bls_sigverifier
that we are doing e.g. in
https://github.com/anza-xyz/alpenglow/pull/582.